fix/перенес типы в одно место, поправил стили
