+++ src/pages/Home/types.tsx
@@ -1,5 +1,7 @@
 export type StatusType = 'queue' | 'progress' | 'done'
 
+export const statuses: StatusType[] = ['queue', 'progress', 'done']
+
 export interface StatusProps {
   status: StatusType
 }
@@ -9,4 +11,18 @@ export interface InlineStatusProps {
   onChange: (status: StatusType) => void
 }
 
-export const statuses: StatusType[] = ['queue', 'progress', 'done']
+export interface BookingRequest {
+  id: number
+  date: string
+  account: string
+  clientName: string
+  service: string
+  projectName: string
+  status: StatusType
+  [key: string]: string | number
+}
+
+export interface FilterDropdownProps {
+  className?: string
+  onChange: (value: string) => void
+}
